pyproject.toml
[project]
name = "NetworkExpertAgent"
version = "0.2.0"
description = "Network Expert Agent with RAG and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "httpx",
    "chromadb",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-text-splitters",
    "langchain-huggingface",
    "sentence-transformers",
    "python-dotenv",
    "langgraph",
    "langsmith",
    "langchain-chroma>=1.1.0",
]

[project.scripts]
network-expert = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
